Modify destroy command to check project directory
Updated the destroy command to check for the docker_build_DHI_LAMP_Project directory before executing docker commands.

+++ main.tf
@@ -31,10 +31,10 @@ resource "null_resource" "lamp_stack" {
     when    = destroy
     command = <<EOT
       # Remove contrainers
-      docker compose down || docker-compose down
-      docker rmi $(docker images --filter "reference=lamp-*" -q)
-      # Remove docker network
-      docker network ls --filter name=lamp_network -q |xargs -r docker network rm
+      if [ -d "docker_build_DHI_LAMP_Project" ]; then
+        cd "docker_build_DHI_LAMP_Project"
+        docker compose down || docker-compose down
+      fi
     EOT
   }
 }
